nlpfinal/exam_evaluator/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
streamlit>=1.33,<2
torch>=2.2,<3
transformers>=4.39,<5
sentence-transformers>=2.6,<4
pillow>=10,<12
